split functionally by module
