+++ elektra/exceptions.py
@@ -1,0 +1,10 @@
+class InsufficientDataError(Exception):
+    pass
+
+
+class ElektraConfigError(Exception):
+    pass
+
+
+class NoRelevantHoursTodayError(Exception):
+    pass
